Prevent overflow in intermediate calculation

+++ radio/src/telemetry/telemetry_sensors.cpp
@@ -353,11 +353,11 @@ void TelemetryItem::eval(const TelemetrySensor & sensor)
           }
         }
         uint32_t angle = abs(gpsItem.gps.latitude - gpsItem.pilotLatitude);
-        uint32_t dist = EARTH_RADIUS * angle / 1000000;
+        uint32_t dist = ((uint64_t)EARTH_RADIUS * angle) / 1000000;
         uint32_t result = dist*dist;
 
         angle = abs(gpsItem.gps.longitude - gpsItem.pilotLongitude);
-        dist = gpsItem.distFromEarthAxis * angle / 1000000;
+        dist = ((uint64_t)gpsItem.distFromEarthAxis * angle) / 1000000;
         result += dist*dist;
 
         // Length on ground (ignoring curvature of the earth)
